Fix for issue #47
Revert procname changes to avoid messing with sysfs

turns out that module_param will, of course!, create:
	/sys/module/<name>/parameters/
forcing kovid to handle the new sysfs kobject and adding
more sysfs handling for init and deinit, something that I
want to avoid.

This way here we still avoid detection via old /proc/kovid,
making the user to explicitly defining a /proc/<name>
for the lkm. So it is up to the user to be creative
and chose a difficult to guess name, hopefully, random, for
his own good.