Hide kovid word from /proc /sys /var/log
This is tricky!

kovid can appear in log files if something goes wrong.

Like taint messages in dmesg, just to name one.

This patch if focused when compiling kovid in release mode:
        DEPLOY=1 make

This way, kv is not supposed to make its way into logs like kern.log
(dmesg) but if that happens kern.log will become "empty" and
current sessions in polling like "dmesg -w" will die (without an error)
and kovid taint messages will not be seen. This a more radical approach
but IMO necessary because the last thing we want is that this lkm
to be found.

I only tackle /proc, /sys and /var/log and let other places
untouched, as expected, one should not include this lkm name
anywhere else.

TODO: make sure magic word does not, by any chance, has "kovid" name
in it ;)