Add generic system command function
Also add -j for clearing journal, currently using journalctl
set in Makefile

Note: this is not perfect solution (as is not dmesg one) it is
a high-level call, but has practicality due to being embedded in the
rootkit.

Unfortunately, and depending on the systemd configuration, journald
will send logs to kernel's ringbuffer, for that reason I also include
"journald" in the set of banned words.

If it goes on dmesg, dmesg call will return 0, far from ideal
behaviour but still less worse than kovid so explicitly visible.

We'll go back at this at some point.