stat: Now decrementing nlinks
Added proc command hide-directory=<dirname>

At filldir fn and statx syscall it is not
straightforward to know wheather a "filename"
is a regular file or a directory.

For us to decrement nlink kstat member, without
knowing if we're handling a directory, I had to, at
the moment the directory is hidden by the user check
dentry with S_IFDIR, and set a boolean.

This boolean is then stored and in statx syscall hook
I search for the name matching inode and directory boolen.

Then nlink can be decremented.