proc: Fix kernel crash related to PROCNAME
This may not be the ideal fix, and further verification is needed. However,
it serves as a temporary solution for now.

Block lseek on /proc/PROCNAME to prevent potential issues.
The way I interact with the proc UI to hide it may cause problems if someone tries to lseek the file,
as I rely on statx and other methods. Blocking lseek is a safer approach for the time being.