back-door: Fix timing issue causing ugly crash sometimes
Replace msleep by a busy loop that can count
the back-door's child tasks. Return actual number,
or 0 if timeout expires, and generate a warning
in DEBUG mode.

For now, don't enforce the minimal number, it may vary, but
it is useful for debugging back-doors, which can be difficult.

Also, fix some variable names and other minor stuff.